+++ CHANGELOG.md
@@ -8,7 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [UNRELEASED]
 
 ### Added
-
+- Added new EnrichRdsFindingAction, which attempts to grab as much details as possible for the instance.
+  - Added unit tests for this new action
+  - Added to the integration tests for RDS actions.
+- Added RdsEnrichmentData Pydantic model to enforce schema.
 
 
 ## [0.10.0] - 2025-10-19